Fixed issue where small blog posts where always 0 minute reading time

+++ src/LinkDotNet.Blog.Web/Features/Services/ReadingTimeCalculator.cs
@@ -13,7 +13,7 @@ public static partial class ReadingTimeCalculator
         var imageCount = ImageRegex().Matches(content).Count;
 
         var wordCount = GetWordCount(content);
-        var readTimeWords = wordCount / wordsPerMinute;
+        var readTimeWords = (double)wordCount / wordsPerMinute;
         var readTimeImages = imageCount * minutesPerImage;
         return (int)Math.Ceiling(readTimeWords + readTimeImages);
     }
